# proc/native: optimize native.status through buffering